Consolidate chat viewmodel/observer code
